% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rastertools.r
\name{calculate_raster_difference}
\alias{calculate_raster_difference}
\title{Raster Difference}
\usage{
calculate_raster_difference(raster1, raster2)
}
\arguments{
\item{raster1}{A \code{SpatRaster} object.}

\item{raster2}{A \code{SpatRaster} object.}
}
\value{
The difference raster, i.e., \code{raster1 - raster2}.
}
\description{
Calculate the difference of two rasters.
}
\keyword{internal}
