% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rastertools.r
\name{calculate_raster_product}
\alias{calculate_raster_product}
\title{Raster Product}
\usage{
calculate_raster_product(raster1, raster2)
}
\arguments{
\item{raster1}{A \code{SpatRaster} object.}

\item{raster2}{A \code{SpatRaster} object.}
}
\value{
The product raster, i.e., \code{raster1 * raster2}.
}
\description{
Calculate the product of two rasters.
}
\keyword{internal}
