% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fema.r
\name{calculate_surcharge}
\alias{calculate_surcharge}
\title{Calculate Surcharge}
\usage{
calculate_surcharge(bfe, floodway_wse, model_elements)
}
\arguments{
\item{bfe}{Base Flood Elevation \code{SpatRaster}.}

\item{floodway_wse}{Floodway water surface elevation \code{SpatRaster}.}

\item{model_elements}{Model elements \code{SpatVector}.}
}
\value{
A classified \code{SpatRaster} object.
}
\description{
Calculate and classify surcharge according to FEMA guidance.
}
