% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fema.r
\name{calculate_unit_discharge}
\alias{calculate_unit_discharge}
\title{Calculate Unit Discharge}
\usage{
calculate_unit_discharge(floodway_dv, model_elements)
}
\arguments{
\item{floodway_dv}{Floodway depth-velocity \code{SpatRaster}.}

\item{model_elements}{Model elements \code{SpatVector}.}
}
\description{
Calculate unit discharge according to FEMA guidance.
Essentially, this function averages the floodway depth x velocity
raster across model elements.
}
