% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rastertools.r
\name{classify_surcharge}
\alias{classify_surcharge}
\title{Classify Surcharge}
\usage{
classify_surcharge(bfe, floodway_wse)
}
\value{
A classified \code{SpatRaster} object.
}
\description{
Classify surcharge according to FEMA guidance.
}
