% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.r
\name{compute_weighted_averages}
\alias{compute_weighted_averages}
\title{Compute Weighted Averages}
\usage{
compute_weighted_averages(numerator, denominator, id)
}
\arguments{
\item{numerator}{The table to use as the numerator.}

\item{denominator}{The table to use as the denominator.}

\item{id}{The field name to group by.}
}
\value{
A data frame.
}
\description{
Computed weighted averages from two tables.
}
\keyword{internal}
