% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fema.r
\name{count_surcharge_exceedance}
\alias{count_surcharge_exceedance}
\title{Count Cells By Surcharge Exceedance}
\usage{
count_surcharge_exceedance(classes)
}
\arguments{
\item{classes}{A vector of surcharge classes.}
}
\value{
A named vector.
}
\description{
Count the number of model elements that exceed FEMA guidance on
allowable surcharge.
}
