% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rastertools.r
\name{evaluate_profiles}
\alias{evaluate_profiles}
\title{Evaluate Profiles}
\usage{
evaluate_profiles(bfe, floodway_wse, floodway_dv, profiles, id = "Name")
}
\arguments{
\item{bfe}{Base Flood Elevation \code{SpatRaster}.}

\item{floodway_wse}{Floodway water surface elevation \code{SpatRaster}.}

\item{floodway_dv}{Floodway depth-velocity \code{SpatRaster}.}

\item{profiles}{Evaluation profiles \code{SpatVector}}
}
\value{
A dataframe.
}
\description{
Evaluate profiles according to FEMA methodology.
}
