% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.r
\name{export_results}
\alias{export_results}
\title{Export Results}
\usage{
export_results(surcharge, evaluation_lines, zipfile = NULL)
}
\arguments{
\item{surcharge}{The surcharge \code{SpatVector}, i.e.,
output of \code{\link[=calculate_surcharge]{calculate_surcharge()}}.}

\item{evaluation_lines}{The evaluation lines \code{SpatVector}, i.e.,
output of \code{\link[=evaluate_profiles]{evaluate_profiles()}}.}

\item{zipfile}{The output file path. If \code{NULL}, a temporary file
will be created. If the file exists, it will be overwritten.}
}
\value{
The output filepath.
}
\description{
Export results to a zip file.
}
