% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.r
\name{extract_cells_by_shape}
\alias{extract_cells_by_shape}
\title{Extract Raster Cells}
\usage{
extract_cells_by_shape(raster, shape, method = c("average", "centroid"))
}
\arguments{
\item{raster}{A file path to a raster.}

\item{shape}{A \code{SpatVector} object or file path to a vector layer.}

\item{method}{The extraction method to use. For
\code{method = "average"}, the average of raster values overlapping
each shape is extracted. For \code{method = "centroid"}, the raster
value occurring at the shape centroids is extracted.}
}
\value{
A subset of \code{raster} cells intersecting with \code{shapes}.
}
\description{
Extract raster cells along a line or shape.
}
\keyword{internal}
