% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.r
\name{get_raster_extent}
\alias{get_raster_extent}
\title{Get Raster Extent}
\usage{
get_raster_extent(raster)
}
\arguments{
\item{raster}{A file path to a raster.}
}
\value{
A \code{SpatExtent} object.
}
\description{
Get the extent from a \code{SpatRaster} object.
}
