% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.r
\name{load_raster_into_memory}
\alias{load_raster_into_memory}
\title{Load Raster}
\usage{
load_raster_into_memory(raster, clip = NULL)
}
\arguments{
\item{raster}{A file path to a raster.}

\item{clip}{A \code{SpatRaster} or \code{SpatExtent} object to crop \code{raster} to.}
}
\value{
A \code{SpatRaster} object.
}
\description{
Load a raster from a file.
}
