% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.r
\name{load_shape_into_memory}
\alias{load_shape_into_memory}
\title{Load Shape}
\usage{
load_shape_into_memory(shape)
}
\arguments{
\item{shape}{A file path to a vector.}
}
\value{
A \code{SpatVector} object.
}
\description{
Load a vector shape from a file.
}
\details{
If \code{shape} is already a \code{SpatVector} object,
the function simply returns the object.
}
