% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.r
\name{load_shapefile}
\alias{load_shapefile}
\title{Load Shapefile}
\usage{
load_shapefile(paths)
}
\arguments{
\item{paths}{A vector of filepaths. Each entry must have a
differentfile  extension.}
}
\value{
A \code{SpatVector} object.
}
\description{
Load a shapefile from a set of (misnamed) files.
}
\keyword{internal}
