% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.r
\name{map_results}
\alias{map_results}
\title{Map Results}
\usage{
map_results(
  surcharge,
  evaluation_lines,
  model_elements,
  bfe,
  floodway_wse,
  floodway_dv
)
}
\arguments{
\item{surcharge}{The surcharge \code{SpatVector}, i.e.,
output of \code{\link[=calculate_surcharge]{calculate_surcharge()}}.}

\item{evaluation_lines}{The evaluation lines \code{SpatVector}, i.e.,
output of \code{\link[=evaluate_profiles]{evaluate_profiles()}}.}

\item{model_elements}{Model elements \code{SpatVector}.}

\item{bfe}{Base Flood Elevation \code{SpatRaster}.}

\item{floodway_wse}{Floodway water surface elevation \code{SpatRaster}.}

\item{floodway_dv}{Floodway depth-velocity \code{SpatRaster}.}
}
\value{
A \code{leaflet} object.
}
\description{
Generate a \code{leaflet} map of results.
}
