% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.r
\name{number_label}
\alias{number_label}
\title{Number Label}
\usage{
number_label(x, digits = 0)
}
\arguments{
\item{x}{A numeric vector.}

\item{digits}{The number of digits to round and display.}
}
\value{
A character vector.
}
\description{
Format numbers for map label display.
}
\keyword{internal}
