% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.r
\name{run_cli}
\alias{run_cli}
\title{Run CLI}
\usage{
run_cli(bfe, floodway_wse, floodway_dv, profiles, model_elements, map = FALSE)
}
\arguments{
\item{bfe}{Base Flood Elevation \code{SpatRaster}.}

\item{floodway_wse}{Floodway water surface elevation \code{SpatRaster}.}

\item{floodway_dv}{Floodway depth-velocity \code{SpatRaster}.}

\item{profiles}{Evaluation profiles \code{SpatVector}.}

\item{model_elements}{Model elements \code{SpatVector}.}

\item{map}{If \code{TRUE}, render an interactive map.}
}
\value{
A named list of results.
}
\description{
Run the Floodway Evaluation command line interface.
}
