% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rastertools.r
\name{count_surcharge_exceedance}
\alias{count_surcharge_exceedance}
\title{Count Cells By Surcharge Exceedance}
\usage{
count_surcharge_exceedance(raster)
}
\value{
A named vector.
}
\description{
Count the number of raster cells that exceed FEMA guidance.
}
