% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rastertools.r
\name{extract_cells_by_shape}
\alias{extract_cells_by_shape}
\title{Extract Raster Cells}
\usage{
extract_cells_by_shape(raster, shapes, id = NULL)
}
\arguments{
\item{raster}{A \code{SpatRaster} object.}

\item{shapes}{A \code{SpatVector} object.}

\item{id}{The attribute of \code{shapes} to use as the ID in the output.}
}
\value{
A subset of \code{raster} cells intersecting with \code{shapes}.
}
\description{
Extract raster cells along a line or shape.
}
\keyword{internal}
