% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.r
\name{map_raster}
\alias{map_raster}
\title{Raster Leaflet Map}
\usage{
map_raster(rastr = NULL, shp = NULL)
}
\arguments{
\item{raster}{A \code{SpatRaster} object.}
}
\value{
A \code{leaflet} object.
}
\description{
Create a leaflet map from a raster.
}
\keyword{internal}
